Major rewrite of xdbg
The original had a number of features for hooking into hydrogen via
comms. However the comms proved to be unreliable in the face of
disconnects/reconnects, and the integration itself worked poorly when
kernels were on a remote machine. It is now removed until I can do a
better job of reimplementing it from scratch.

The other big change is that I've figured out how to do bytecode
manipulation to set breakpoints after a function has already been
constructed. This allows endowing xdbg with more features of a proper
debugger. As a tradeoff no-argument inline %break is removed for now,
but there is a sketch of how to bring it back in the future.